Re-enable staticfilesystem on Travis
This was broken because of `--gc-sections`, although the cause wasn't
quite obvious (and I still don't totally understand it): referencing the
file with `printk_helper()` but not `printk_helper()` itself caused
module symbol versions for `printk` and `__fentry__` to get dropped. #20
changed around linking, causing the entire file to get GC'd unless you
referenced `bug_helper()`, in which case `__fentry__` was fine but
`printk` was not. Dropping `--gc-sections` in #28 actually avoids the
whole problem.