Remove kernel-cflags-finder (fixes #191)
The bulk of this is the changes from #185 that didn't make it into #188.

We also upgrade Travis's Clang/LLVM version to 9 to pick up some sort
of bugfix in relocation processing. With older versions, we get

[ 1827.072410] module: x86/modules: Skipping invalid relocation target, existing value is nonzero for type 1, loc 00000000197e607d, val ffffffffc0703520

(where that "loc" seems to be random/garbage data, disconcertingly).
Clang 9 is already required by kernels 5.0+ so this shouldn't be too
much of a problem.

Co-authored-by: Alex Gaynor <alex.gaynor@gmail.com>